Create pages for each of the Application component
The architecture of the whole application system is based on the
navigation between separate components implemented in the React Native.
These components are to be rendered as the UWP's Page content, just like
in the original RNW application.
The difference is, that in this approach each RN Component will be
rendered inside separate Page as it's own ReactNative Root View.

This approach requires all Components to have the separate Page
implemented, each containing XAML scheme, IDL implementation and it's
code-behind implemented in native C++.