Create main panel for notes
The application's main panel contains all the notes added by the user.
They are displayed in the grid, with number of columns adapted to the
width of the window.
Note, that the Window size needs to be considered instead of Screen's.
That is due to having the Page in RNW rendered as a root with certain
and fixed dimensions.

Currently only the testing data are given to the FlatList, so there's a
possibility to test it by displaying any content.
In further development that data should be taken from the database.