Implement navigation native module
This native module is responsible for handling the navigation between
screens which are not connected through the NavigationView and it's
Frame.
This situation happens, when there's a need to navigate to separate
screen, but at the same time get rid of navigation panel.

This native module is called  NoteWidgetClickHandler  because it is
designed explicitly for the  NotesPage <-> NoteWidgetDetailsPage
navigation both-ways.