Implement going back function for NoteWidgetClickHandler
The native module should also provide the system with the ability to go
back.
This is now done by the GoToNotesScreen function, which navigates to the
MainPage.

NOTE: MainPage needs to be selected, as only this page contains the
NavigationFrame and is able to display the navigation top pane.
This is why there's a starting point implemented when MainPage is
navigated to.