Create in-memory repository for notes storage
The Repository component is responsible for storing all the data in the
random access memory of the application.
This solution is to avoid the overkill of adding and managing the
database engine when implementing application as MVP.

The Repository class has four main methods allowing to manipulate the
container of notes in the systen: Create, Read, Update, Delete (CRUD).
Each method implements operations on the std::vector containing the
NoteModel's entities.
This repository is created as a Singleton, to utilize the following
advantages of it:
* It's kept global, so the vector of NoteModel will never be destroyed
  unless application is closed.
* It's accessible by any other component in the C++/WinRT layer
* It will not be destroyed between each JS native module call.
  Such case would be possible as each call to native module or native UI
  component creates new instance of it, so Repository and it's container
  would be destroyed.
  Singleton pattern used here prevents from such scenario.