Create Database handler
The handler for the database now operates on the Repository pattern,
which manipulates the in-memory data.
It is the native module, which gives the ability to add or get a note
from the application.
The note returned by this native module is returned property-by-property
because there's no possibility to return the whole object to the JS
side.