Create notes dependning on the number of created notes
This implementation adds several methods to the Database handler class:
* GetNumberOfNotes - returns the number of notes already created by the
  user in the system. It considers only the size of the repository,
  without checking the correctness of the IDs.
* DoesIDExists - checks whether the notes with the given ID exists in
  the Database
* GetAllNotes - alternate version of creating the notes keys' method.
  It returns the array of already created objects containing the Key and
  it's value.
  This is experimental and not yet fully tested/supported.

Note that the `GetNumberOfNotes` method uses Promise instead of return
callback. This is due to the fact of async nature of native modules. So
to avoid losing the information of returned value when going out of the
return handler, the promise is used which carries the result.