Overload the CRUD operations for index
The index can now be used as a parameter for all the CRUD operations
done on the Repository.
This is to allow the easier manipulations when it comes to all notes no
matter if they Id matches the db position.

The purpose of this implementation is to make sure that in the future
the index and the ID will be separate and the logic behind
adding/removing and displaying the content will be more flexible.