Show created notes on the main page (#7)
* Create notes dependning on the number of created notes

This implementation adds several methods to the Database handler class:
* GetNumberOfNotes - returns the number of notes already created by the
  user in the system. It considers only the size of the repository,
  without checking the correctness of the IDs.
* DoesIDExists - checks whether the notes with the given ID exists in
  the Database
* GetAllNotes - alternate version of creating the notes keys' method.
  It returns the array of already created objects containing the Key and
  it's value.
  This is experimental and not yet fully tested/supported.

Note that the `GetNumberOfNotes` method uses Promise instead of return
callback. This is due to the fact of async nature of native modules. So
to avoid losing the information of returned value when going out of the
return handler, the promise is used which carries the result.

* Go the main page when entering the app

* Overload the CRUD operations for index

The index can now be used as a parameter for all the CRUD operations
done on the Repository.
This is to allow the easier manipulations when it comes to all notes no
matter if they Id matches the db position.

The purpose of this implementation is to make sure that in the future
the index and the ID will be separate and the logic behind
adding/removing and displaying the content will be more flexible.

* Use Promise instead of callback for the Database

The Database native module has all the methods returning values used
asynchronously, but their results are considered as states on the JS
side.
To avoid loosing the data the Promise has been used to make the database
pull more dynamic.

* Pull title and short message when creating note widget

* Use Hooks in NoteWidget