Fix: memory leak when resizing the app window
The resizing of the window app has a memory leak caused by the
asynchronous calls still running on the unmounted components (which
applies to both Note main panel and note widget).
To fix that, the simple mounted status has been added which is checked
each time the component is about to be updated, so once the component is
unmounted the rerender will not happen.