Make the TitleBar transparent
The Title bar of the application is the most default app element that is
shown to the user, that is why, to make it more custom and to make the
application look more modern, the title bar was made transparent.
The transparency means that:
* the content of the app is extended to the full size of the whole
  window,
* the color of the title bar is set to transparent,
* the background color of the window manipulation buttons (maximize,
  minimize, close) is set to transparent
  It makes them still being visible and fully operative, but are blended
  into the app window

NOTE: The title bar has still it's default size, so to drag the window,
or double click, it's still possible to manipulate the title bar as it
would still be there.