Add gradient to the application's page
Each page of the application has been provided with the gradient set as
a background.
This feature is implemented on the UWP side due to no decent support for
the gradient on the React Native side (especially no support for Windows).

The gradient is set as the direct child of the Background property of
the ReactRootView component of the XAML RNW element.