Catch the exception in case of general window failure
When trying to open the window on the separate thread (after launching
the coroutine) it is possible that there'll be the exception thrown by
the base.h implementation.
The reason for specifying the winrt::hresult_error is that it does not
inherits from the std::exception so it would be impossible to catch by
the universal exception.