Fix: Invalid window handle due when opening FilePicker
The FilePicker when opened was throwing the "Invalid window handle" error.
This was caused by the native module thread affinity and the fact that
the FilePicker should be launched in the UI thread explicitly.

The fix for that is to launch the couroutine on the UI context using the
UIDispatcher method of React Native Windows.
for more details and the reasoning please check:
https://github.com/microsoft/react-native-windows/issues/7641