Simplify the Notes FlatList handling
The FlatList containing all the notes created in the application were
previously being rerendered each time the window's size changed.
This has been optimized by handling the whole database fetch procedure
in the main notes screen instead of delegating it into each note widget.

The current solution is that when NotesScreen page loads the database
returns the full set of notes already packed as objects.
This is returned as a Promise, which allows to render each note
separately and only then fill them with data.